# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects/mine": {
      "get": {
        "description": "Retorna todos los proyectos a los que tiene acceso el usuario autenticado",
        "operationId": "ProjectController_findMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de proyectos del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreateProjectDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado - Token inválido o expirado"
          },
          "403": {
            "description": "Acceso denegado - Información de tenant no disponible"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Obtener proyectos del usuario actual",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects": {
      "post": {
        "description": "Creates a new project for the current tenant. Accessible by global admins or tenant admins.",
        "operationId": "ProjectController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data - Check the request body format"
          },
          "401": {
            "description": "Unauthorized - Invalid or expired token"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions to create projects"
          },
          "409": {
            "description": "Project already exists - A project with this name already exists for this tenant"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create new project",
        "tags": [
          "Projects"
        ]
      },
      "get": {
        "description": "Retrieves a list of all projects for the current tenant. Results are cached for 1 hour.",
        "operationId": "ProjectController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of projects retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or expired token"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Get all projects",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/{id}": {
      "get": {
        "description": "Retrieves a specific project by its unique ID. Results are cached for 1 hour.",
        "operationId": "ProjectController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique project identifier (UUID)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or expired token"
          },
          "404": {
            "description": "Project not found - The specified ID does not exist for this tenant"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get project by ID",
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "description": "Updates an existing project's information. Accessible by global admins or tenant admins.",
        "operationId": "ProjectController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique project identifier to update (UUID)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data - Check the request body format"
          },
          "401": {
            "description": "Unauthorized - Invalid or expired token"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions to update projects"
          },
          "404": {
            "description": "Project not found - The specified ID does not exist for this tenant"
          },
          "409": {
            "description": "Project name already exists - The provided name is already in use"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update project",
        "tags": [
          "Projects"
        ]
      },
      "delete": {
        "description": "Permanently deletes a project. This is a destructive operation that requires admin privileges.",
        "operationId": "ProjectController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique project identifier to delete (UUID)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Project successfully deleted"
          },
          "401": {
            "description": "Unauthorized - Invalid or expired token"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions to delete projects"
          },
          "404": {
            "description": "Project not found - The specified ID does not exist for this tenant"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete project",
        "tags": [
          "Projects"
        ]
      }
    },
...
  },
  "info": {
    "title": "japm.app API",
    "description": "API for the japm.app Prompt Engineering application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the project",
            "example": "My Awesome Project"
          },
          "description": {
            "type": "string",
            "description": "Optional description for the project",
            "example": "A project to demonstrate NestJS and Prisma"
          }
        },
        "required": [
          "name"
        ]
      },
      "ProjectDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the project",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Name of the project",
            "example": "My Project"
          },
          "description": {
            "type": "string",
            "description": "Optional description of the project",
            "example": "A project for managing content"
          },
          "tenantId": {
            "type": "string",
            "description": "ID of the tenant this project belongs to",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "ownerUserId": {
            "type": "string",
            "description": "ID of the user who owns this project",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp",
            "example": "2024-01-01T00:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp",
            "example": "2024-01-01T00:00:00Z"
          }
        },
        "required": [
          "id",
          "name",
          "tenantId",
          "ownerUserId",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the project",
            "example": "My Awesome Project"
          },
          "description": {
            "type": "string",
            "description": "Optional description for the project",
            "example": "A project to demonstrate NestJS and Prisma"
          }
        }
      },
...
    }
  }
}